Reorganize the packages of input,output and strem
