+++ src/main/java/com/intel/chimera/stream/input/CTRCryptoInputStream.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.chimera.stream;
+package com.intel.chimera.stream.input;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -31,9 +31,6 @@ import javax.crypto.ShortBufferException;
 
 import com.intel.chimera.cipher.Cipher;
 import com.intel.chimera.cipher.CipherTransformation;
-import com.intel.chimera.input.ChannelInput;
-import com.intel.chimera.input.Input;
-import com.intel.chimera.input.StreamInput;
 import com.intel.chimera.utils.Utils;
 
 /**
