+++ src/main/java/com/intel/chimera/stream/input/ChannelInput.java
@@ -15,55 +15,59 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.chimera.input;
+package com.intel.chimera.stream.input;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.ByteBuffer;
+import java.nio.channels.ReadableByteChannel;
 
 /**
- * The StreamInput class takes a <code>InputStream</code> object and
+ * The ChannelInput class takes a <code>ReadableByteChannel</code> object and
  * wraps it as <code>Input</code> object acceptable by <code>CryptoInputStream</code>.
  */
-public class StreamInput implements Input {
-  private byte[] buf;
-  private int bufferSize;
-  InputStream in;
+public class ChannelInput implements Input {
+  private static final int SKIP_BUFFER_SIZE = 2048;
 
-  public StreamInput(InputStream inputStream, int bufferSize) {
-    this.in = inputStream;
-    this.bufferSize = bufferSize;
+  private ByteBuffer buf;
+  private ReadableByteChannel channel;
+
+  public ChannelInput(
+      ReadableByteChannel channel) {
+    this.channel = channel;
   }
 
   @Override
   public int read(ByteBuffer dst) throws IOException {
-    int remaining = dst.remaining();
-    final byte[] tmp = getBuf();
-    int read = 0;
-    while (remaining > 0) {
-      final int n = in.read(tmp, 0, Math.min(remaining, bufferSize));
-      if (n == -1) {
-        if (read == 0) {
-          read = -1;
-        }
-        break;
-      } else if (n > 0) {
-        dst.put(tmp, 0, n);
-        read += n;
-        remaining -= n;
-      }
-    }
-    return read;
+    return channel.read(dst);
   }
 
   @Override
   public long skip(long n) throws IOException {
-    return in.skip(n);
+    long remaining = n;
+    int nr;
+
+    if (n <= 0) {
+      return 0;
+    }
+
+    int size = (int)Math.min(SKIP_BUFFER_SIZE, remaining);
+    ByteBuffer skipBuffer = getSkipBuf();
+    while (remaining > 0) {
+      skipBuffer.clear();
+      skipBuffer.limit((int)Math.min(size, remaining));
+      nr = read(skipBuffer);
+      if (nr < 0) {
+        break;
+      }
+      remaining -= nr;
+    }
+
+    return n - remaining;
   }
 
   @Override
   public int available() throws IOException {
-    return in.available();
+    return 0;
   }
 
   @Override
@@ -81,12 +85,12 @@ public class StreamInput implements Input {
 
   @Override
   public void close() throws IOException {
-    in.close();
+    channel.close();
   }
 
-  private byte[] getBuf() {
+  private ByteBuffer getSkipBuf() {
     if (buf == null) {
-      buf = new byte[bufferSize];
+      buf = ByteBuffer.allocate(SKIP_BUFFER_SIZE);
     }
     return buf;
   }
