+++ src/main/java/com/intel/chimera/stream/output/CTRCryptoOutputStream.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.chimera.stream;
+package com.intel.chimera.stream.output;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -31,9 +31,6 @@ import javax.crypto.ShortBufferException;
 
 import com.intel.chimera.cipher.Cipher;
 import com.intel.chimera.cipher.CipherTransformation;
-import com.intel.chimera.output.ChannelOutput;
-import com.intel.chimera.output.Output;
-import com.intel.chimera.output.StreamOutput;
 import com.intel.chimera.utils.Utils;
 
 /**
